tested serial flash read back

+++ HAL/memory/m25p64/m25p64.hpp
@@ -30,7 +30,7 @@ public:
     std::int16_t program_page(std::uint32_t, const void *, std::uint8_t);
     std::int16_t verify_electonic_signature(void);
     std::int16_t verify_identification(void);
-    std::int16_t read(std::uint32_t, void *, std::uint32_t);
+    std::int16_t read(std::uint32_t, void *, std::size_t);
 
 private:
     bsp::spi &rspi;
